movistar_epg.py & vod.py: stop retrying incomplete recordings after 3 consecutive attempts with the same recorded time.

+++ vod.py
@@ -164,11 +164,16 @@ def ffmpeg_completed():
 
         proc.wait()
 
-    if check_recording():
+    missing_time = check_recording()
+    if not missing_time:
         save_metadata()
         httpx.put(epg_url)
     elif os.path.exists(filename + VID_EXT):
-        os.remove(filename + VID_EXT)
+        resp = httpx.put(epg_url + f"?missing={missing_time}")
+        if resp.status_code == 200:
+            save_metadata()
+        else:
+            os.remove(filename + VID_EXT)
 
     _cleanup()
 
@@ -206,7 +211,7 @@ def _handle_cleanup(signum, frame):
 def check_recording():
     if not os.path.exists(filename + VID_EXT):
         sys.stderr.write(f"{_log_prefix} Recording CANNOT FIND: {_log_suffix}")
-        return False
+        return -1
 
     command = list(_nice)
     command += ["ffprobe", "-v", "quiet", "-print_format", "json", "-show_format"]
@@ -216,7 +221,7 @@ def check_recording():
         duration = int(float(probe["format"]["duration"]))
     except (KeyError, ValueError):
         sys.stderr.write(f"{_log_prefix} Recording CANNOT PARSE: {_log_suffix}")
-        return False
+        return -1
 
     _bad = (duration + 30) < args.time
     sys.stderr.write(
@@ -225,7 +230,7 @@ def check_recording():
         f"{_log_suffix}"
     )
 
-    return not _bad
+    return 0 if not _bad else args.time - duration
 
 
 def find_free_port(iface=""):
